+++ packages/instant-apps-components/src/components/instant-apps-language-translator/support/utils.ts
@@ -93,3 +93,14 @@ export function initExternalCKEditorStyles() {
   style.innerHTML = CKEditorStyles;
   document.head.appendChild(style);
 }
+
+export async function updateLastSave(resource: __esri.PortalItemResource): Promise<void> {
+  const data = await resource.fetch();
+  const lastSave = Date.now();
+  store.set('lastSave', lastSave);
+  const dataStr = JSON.stringify({ ...data, lastSave });
+  const blobParts = [dataStr];
+  const options = { type: 'application/json' };
+  const blob = new Blob(blobParts, options);
+  resource.update(blob);
+}